+++ R/func_gg.R
@@ -3,7 +3,7 @@
 # Use      : Wrapper functions for ggplot2
 # Author   : Tomas Sou
 # Created  : 2025-10-25
-# Updated  : 2026-03-03
+# Updated  : 2026-03-13
 #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 # Notes
 # na
@@ -226,20 +226,20 @@ ggvio = function(d, var, cats, alpha=0.1, show=TRUE, nsub=TRUE, ...){
 #' XY scatter plot wrapper
 #'
 #' Create basic XY scatter plot for quick data exploration.
-#' Default to show Pearson correlation coefficient with p-value using [ggpubr::stat_cor].
+#' Default to show Pearson correlation coefficient with p-value using [ggpubr::stat_cor()].
 #' For more complex plots, it is recommended to use [ggplot2] directly.
 #'
 #' @param d `<dfr>` A data frame.
 #' @param x,y `<var>` Variables for x- and y-axis as unquoted names.
-#' @param ... Arguments to pass to [ggplot2::aes] for additional mapping.
+#' @param ... Arguments to pass to [ggplot2::aes()] for additional mapping.
 #' @param lm `<lgl>` `TRUE` to add regression line from linear model.
 #' @param se `<lgl>` `TRUE` to show standard error with the regression line.
 #' @param nsub `<lgl>` Show number of observations.
 #' @param cor `<lgl>` `TRUE` to show Pearson correlation coefficient with p-value.
 #' @param pv `<dbl>` Precision of p-value, e.g., 0.001 to show 3 decimal places, `NULL` for no rounding.
 #' @param legend `<lgl>` `TRUE` to show legend.
-#' @param asp `<num>` For `aspect.ratio` in [ggplot2::theme].
-#' @seealso [ggpubr::stat_cor]
+#' @param asp `<num>` For `aspect.ratio` in [ggplot2::theme()].
+#' @seealso [ggpubr::stat_cor()]
 #' @returns A ggplot object.
 #' @export
 #' @examples